+++ app/src/main/java/com/wepindia/pos/BillingDineInActivity.java
@@ -4999,7 +4999,7 @@ public class BillingDineInActivity extends WepPrinterBaseActivity {
             Double qty = Double.parseDouble(itemQty.getText().toString().trim());
             double rate = Double.parseDouble(itemRate.getText().toString().trim());
             double amount = Double.parseDouble(itemAmount.getText().toString().trim());
-            BillKotItem billKotItem = new BillKotItem(sno, name, qty.intValue(), rate, amount);
+            BillKotItem billKotItem = new BillKotItem(sno, name, qty, rate, amount);
             billKotItems.add(billKotItem);
             count++;
